* Removed too much abstraction * Decreased the number of files
